[feature] DynamicResourceHandlerInitializer 기능 추가  - @ApiHandler가 붙은 클래스의 생성자에서 @DataHandler가 붙은 클래스 주입
